requirements.txt
dynaconf
numpy==1.26.4
pandas==2.1.4
python-dotenv==1.0.1
openai==1.34.0
ibm_watsonx_ai==1.0.6
langchain-ibm==0.1.7
langchain_community==0.2.4
ragas==0.1.9
unstructured==0.14.5
milvus==2.3.8
pymilvus==2.4.3
sentence-transformers==3.0.1
accelerate==0.33.0
